Adjust CLM frontend namespacing
