--- arches_references/src/arches-references/components/editor/ItemURI.vue
@@ -1,171 +1,0 @@
-<script setup lang="ts">
-import { computed, inject, ref } from "vue";
-import { useGettext } from "vue3-gettext";
-
-import InputText from "primevue/inputtext";
-import { useToast } from "primevue/usetoast";
-
-import { ARCHES_CHROME_BLUE } from "@/arches/theme.ts";
-import { patchItem } from "@/arches-references/api.ts";
-import {
-    DEFAULT_ERROR_TOAST_LIFE,
-    ERROR,
-    itemKey,
-} from "@/arches-references/constants.ts";
-import { vFocus } from "@/arches-references/utils.ts";
-
-import type { Ref } from "vue";
-import type { ControlledListItem } from "@/arches-references/types";
-
-const item = inject(itemKey) as Ref<ControlledListItem>;
-
-const editing = ref(false);
-
-const formValue = ref("");
-
-const inputValue = computed({
-    get() {
-        return item.value.uri;
-    },
-    set(newVal: string) {
-        formValue.value = newVal;
-    },
-});
-
-const toast = useToast();
-const { $gettext } = useGettext();
-const uri = "uri";
-
-const save = async () => {
-    editing.value = false;
-    const originalValue = item.value.uri;
-    item.value.uri = formValue.value;
-
-    try {
-        await patchItem(item.value, uri);
-    } catch (error) {
-        toast.add({
-            severity: ERROR,
-            life: DEFAULT_ERROR_TOAST_LIFE,
-            summary: $gettext("Save failed"),
-            detail: error instanceof Error ? error.message : undefined,
-        });
-        item.value.uri = originalValue;
-    }
-};
-
-const cancel = () => {
-    editing.value = false;
-    formValue.value = item.value.uri;
-};
-</script>
-
-<template>
-    <div class="uri-container">
-        <h4>{{ $gettext("List Item URI") }}</h4>
-        <p>
-            {{
-                $gettext(
-                    "Optionally, provide a URI for your list item. Useful if your list item is formally defined in a thesaurus or authority document.",
-                )
-            }}
-        </p>
-        <div class="characteristic">
-            <InputText
-                v-model="inputValue"
-                v-focus
-                type="text"
-                :disabled="!editing"
-                :aria-label="$gettext('Enter a URI')"
-                :placeholder="$gettext('Enter a URI')"
-                @keyup.enter="save"
-            />
-            <span
-                v-if="!editing"
-                class="edit-controls"
-            >
-                <i
-                    role="button"
-                    tabindex="0"
-                    class="fa fa-pencil"
-                    :aria-label="$gettext('Edit')"
-                    @click="editing = true"
-                    @keyup.enter="editing = true"
-                />
-            </span>
-            <span
-                v-if="editing"
-                class="edit-controls"
-            >
-                <i
-                    role="button"
-                    tabindex="0"
-                    class="fa fa-check"
-                    :aria-label="$gettext('Save edit')"
-                    @click="save"
-                    @keyup.enter="save"
-                />
-                <i
-                    role="button"
-                    tabindex="0"
-                    class="fa fa-undo"
-                    :aria-label="$gettext('Cancel edit')"
-                    @click="cancel"
-                    @keyup.enter="cancel"
-                />
-            </span>
-        </div>
-    </div>
-</template>
-
-<style scoped>
-.uri-container {
-    margin: 1rem 1rem 3rem 1rem;
-    display: flex;
-    flex-direction: column;
-    width: 100%;
-}
-h4 {
-    color: v-bind(ARCHES_CHROME_BLUE);
-    margin-top: 0;
-    font-size: 1.33rem;
-}
-
-p {
-    font-weight: normal;
-    margin-top: 0;
-    font-size: 1.2rem;
-}
-
-input {
-    font-size: 1.2rem;
-}
-
-.characteristic {
-    margin: 1rem;
-    display: flex;
-    align-items: center;
-}
-
-.characteristic input {
-    text-align: center;
-    height: 3rem;
-    width: 100%;
-}
-
-.characteristic input[disabled] {
-    text-align: left;
-}
-
-.edit-controls {
-    margin-left: 1rem;
-    display: inline-flex;
-    justify-content: space-between;
-    width: 4rem;
-}
-
-.edit-controls i {
-    font-size: small;
-    padding: 0.5rem;
-}
-</style>
