--- arches_references/src/arches-references/components/tree/MoveRow.vue
@@ -1,272 +1,0 @@
-<script setup lang="ts">
-import { inject, ref, watch } from "vue";
-import { useGettext } from "vue3-gettext";
-
-import Button from "primevue/button";
-import { useToast } from "primevue/usetoast";
-
-import { patchList } from "@/arches-references/api.ts";
-import {
-    DEFAULT_ERROR_TOAST_LIFE,
-    ERROR,
-    PREF_LABEL,
-    displayedRowKey,
-    selectedLanguageKey,
-} from "@/arches-references/constants.ts";
-import {
-    findNodeInTree,
-    itemAsNode,
-    listAsNode,
-    nodeIsItem,
-    nodeIsList,
-    reorderItems,
-} from "@/arches-references/utils.ts";
-
-import type { Ref } from "vue";
-import type { TreeExpandedKeys, TreeSelectionKeys } from "primevue/tree/Tree";
-import type { TreeNode } from "primevue/treenode";
-import type { Language } from "@/arches/types";
-import type {
-    ControlledList,
-    ControlledListItem,
-    DisplayedListItemRefAndSetter,
-    MoveLabels,
-    NewControlledListItem,
-} from "@/arches-references/types";
-
-const toast = useToast();
-const { $gettext } = useGettext();
-
-const selectedLanguage = inject(selectedLanguageKey) as Ref<Language>;
-const { displayedRow, setDisplayedRow } = inject(
-    displayedRowKey,
-) as DisplayedListItemRefAndSetter;
-
-const { moveLabels, node } = defineProps<{
-    moveLabels: MoveLabels;
-    node: TreeNode;
-}>();
-
-const tree = defineModel<TreeNode[]>("tree", { required: true });
-const expandedKeys = defineModel<TreeExpandedKeys>("expandedKeys", {
-    required: true,
-});
-const selectedKeys = defineModel<TreeSelectionKeys>("selectedKeys", {
-    required: true,
-});
-const movingItem = defineModel<TreeNode>("movingItem", { required: true });
-const nextNewItem = defineModel<NewControlledListItem>("nextNewItem");
-const newLabelFormValue = defineModel<string>("newLabelFormValue", {
-    required: true,
-});
-const newLabelCounter = ref(1);
-const shouldRefocusUpArrow = ref(false);
-const shouldRefocusDownArrow = ref(false);
-
-watch(displayedRow, () => {
-    shouldRefocusUpArrow.value = false;
-    shouldRefocusDownArrow.value = false;
-});
-
-const isFirstItem = (item: ControlledListItem) => {
-    const siblings: TreeNode[] = item.parent_id
-        ? findNodeInTree(tree.value, item.parent_id).data.children
-        : findNodeInTree(tree.value, item.list_id).data.items;
-    if (!siblings) {
-        throw new Error();
-    }
-    return siblings[0].id === item.id;
-};
-
-const isLastItem = (item: ControlledListItem) => {
-    const siblings: TreeNode[] = item.parent_id
-        ? findNodeInTree(tree.value, item.parent_id).data.children
-        : findNodeInTree(tree.value, item.list_id).data.items;
-    if (!siblings) {
-        throw new Error();
-    }
-    return siblings[siblings.length - 1].id === item.id;
-};
-
-const setMovingItem = (node: TreeNode) => {
-    movingItem.value = findNodeInTree(
-        [itemAsNode(displayedRow.value, selectedLanguage.value)],
-        node.key,
-    );
-};
-
-const addItem = (parent: TreeNode) => {
-    const newItem: NewControlledListItem = {
-        parent_id: nodeIsItem(parent) ? parent.data.id : null,
-        id: newLabelCounter.value,
-        list_id: parent.data.list_id ?? parent.data.id,
-        uri: "",
-        sortorder: 0,
-        guide: false,
-        values: [
-            {
-                id: 0,
-                valuetype_id: PREF_LABEL,
-                language_id: selectedLanguage.value.code,
-                value: "",
-                list_item_id: newLabelCounter.value,
-            },
-        ],
-        images: [],
-        children: [],
-        depth: !parent.depth ? 0 : parent.depth + 1,
-    };
-
-    nextNewItem.value = newItem;
-    newLabelFormValue.value = "";
-    newLabelCounter.value += 1;
-
-    parent.children!.push(itemAsNode(newItem, selectedLanguage.value));
-
-    expandedKeys.value = {
-        ...expandedKeys.value,
-        [parent.key as string]: true,
-    };
-    selectedKeys.value = { [newItem.id]: true };
-    setDisplayedRow(newItem);
-};
-
-const reorder = async (item: ControlledListItem, up: boolean) => {
-    const list: ControlledList = findNodeInTree(tree.value, item.list_id).data;
-
-    let siblings: ControlledListItem[];
-    if (item.parent_id) {
-        siblings = findNodeInTree(tree.value, item.parent_id).children.map(
-            (child: TreeNode) => child.data,
-        );
-    } else {
-        siblings = list.items;
-    }
-
-    reorderItems(list, item, siblings, up);
-    const field = "sortorder";
-
-    try {
-        await patchList(list, field);
-    } catch (error) {
-        toast.add({
-            severity: ERROR,
-            life: DEFAULT_ERROR_TOAST_LIFE,
-            summary: $gettext("Save failed"),
-            detail: error instanceof Error ? error.message : undefined,
-        });
-        return;
-    }
-    const oldListIndex = tree.value.findIndex(
-        (listNode) => listNode.data.id === list.id,
-    );
-    tree.value = [
-        ...tree.value.slice(0, oldListIndex),
-        listAsNode(list, selectedLanguage.value),
-        ...tree.value.slice(oldListIndex + 1),
-    ];
-    selectedKeys.value = {
-        ...selectedKeys.value,
-        [item.id]: true,
-    };
-
-    if (up) {
-        shouldRefocusUpArrow.value = true;
-        shouldRefocusDownArrow.value = false;
-    } else {
-        shouldRefocusUpArrow.value = false;
-        shouldRefocusDownArrow.value = true;
-    }
-};
-
-const vRefocusUpArrow = {
-    mounted: (el: HTMLButtonElement) => {
-        if (shouldRefocusUpArrow.value && el) {
-            // @ts-expect-error focusVisible not yet in typeshed
-            el.focus({ focusVisible: true });
-        }
-    },
-};
-const vRefocusDownArrow = {
-    mounted: (el: HTMLButtonElement) => {
-        if (shouldRefocusDownArrow.value && el) {
-            // @ts-expect-error focusVisible not yet in typeshed
-            el.focus({ focusVisible: true });
-        }
-    },
-};
-</script>
-
-<template>
-    <Button
-        v-if="selectedKeys && node.key in selectedKeys"
-        v-tooltip="moveLabels.addChild"
-        type="button"
-        raised
-        class="add-child-button"
-        icon="fa fa-plus"
-        :aria-label="moveLabels.addChild"
-        @click.stop="addItem(node)"
-    />
-    <span
-        v-if="!nodeIsList(node)"
-        class="move-buttons"
-    >
-        <Button
-            v-if="selectedKeys && node.key in selectedKeys"
-            v-refocus-up-arrow
-            v-tooltip="moveLabels.moveUp"
-            type="button"
-            raised
-            class="reorder-button"
-            icon="fa fa-caret-up"
-            :aria-label="moveLabels.moveUp"
-            :disabled="isFirstItem(node.data)"
-            @click="reorder(node.data, true)"
-        />
-        <Button
-            v-if="selectedKeys && node.key in selectedKeys"
-            v-refocus-down-arrow
-            v-tooltip="moveLabels.moveDown"
-            type="button"
-            raised
-            class="reorder-button"
-            icon="fa fa-caret-down"
-            :aria-label="moveLabels.moveDown"
-            :disabled="isLastItem(node.data)"
-            @click="reorder(node.data, false)"
-        />
-        <Button
-            v-if="!node.data.name && selectedKeys && node.key in selectedKeys"
-            v-tooltip="moveLabels.changeParent"
-            type="button"
-            raised
-            icon="fa fa-arrows-alt"
-            :aria-label="moveLabels.changeParent"
-            @click="setMovingItem(node)"
-        />
-    </span>
-</template>
-
-<style scoped>
-.p-button {
-    background-color: aliceblue;
-    color: black;
-    height: 2rem;
-}
-
-.add-child-button {
-    width: 2rem;
-    border-radius: 50%;
-}
-
-.move-buttons {
-    display: flex;
-    gap: 0.5rem;
-    padding-right: 0.5rem;
-}
-
-.move-button {
-    height: 2.5rem;
-}
-</style>
