--- arches_references/src/arches/theme.ts
@@ -1,5 +1,0 @@
-// Temporary place to stash theme constants until PrimeVue 4
-// https://github.com/archesproject/arches/issues/11050
-
-export const ARCHES_CHROME_BLUE = "#2d3c4b";
-export const LIGHT_GRAY = "#f4f4f4";
