+++ arches_references/src/controlled-lists/components/editor/ItemImages.vue
@@ -1,0 +1,144 @@
+<script setup lang="ts">
+import arches from "arches";
+import Cookies from "js-cookie";
+import { inject } from "vue";
+import { useGettext } from "vue3-gettext";
+
+import FileUpload from "primevue/fileupload";
+import { useToast } from "primevue/usetoast";
+
+import { ARCHES_CHROME_BLUE } from "@/arches/theme.ts";
+import {
+    itemKey,
+    DEFAULT_ERROR_TOAST_LIFE,
+    ERROR,
+} from "@/controlled-lists/constants.ts";
+import ImageEditor from "@/controlled-lists/components/editor/ImageEditor.vue";
+
+import type { Ref } from "vue";
+import type { ControlledListItem } from "@/controlled-lists/types";
+import type {
+    FileUploadBeforeSendEvent,
+    FileUploadErrorEvent,
+    FileUploadUploadEvent,
+} from "primevue/fileupload";
+
+const item = inject(itemKey) as Ref<ControlledListItem>;
+
+const { $gettext } = useGettext();
+const toast = useToast();
+
+const addHeader = (event: FileUploadBeforeSendEvent) => {
+    const token = Cookies.get("csrftoken");
+    if (token) {
+        event.xhr.setRequestHeader("X-CSRFToken", token);
+        event.formData.set("list_item_id", item.value!.id);
+    }
+};
+
+const upload = (event: FileUploadUploadEvent) => {
+    if (event.xhr.status !== 201) {
+        showError(event);
+        return;
+    }
+    const newImage = JSON.parse(event.xhr.responseText);
+    item.value!.images.push(newImage);
+};
+
+const showError = (event?: FileUploadErrorEvent | FileUploadUploadEvent) => {
+    toast.add({
+        severity: ERROR,
+        life: DEFAULT_ERROR_TOAST_LIFE,
+        summary: event?.xhr?.statusText || $gettext("Image upload failed"),
+    });
+};
+</script>
+
+<template>
+    <div class="images-container">
+        <h4>{{ $gettext("Images") }}</h4>
+        <FileUpload
+            accept="image/*"
+            :url="arches.urls.controlled_list_item_image_add"
+            :auto="true"
+            :max-file-size="5e6"
+            :file-limit="10"
+            :preview-width="250"
+            :with-credentials="true"
+            :show-cancel-button="false"
+            :show-upload-button="false"
+            name="item_image"
+            :pt="{
+                buttonbar: {
+                    style: {
+                        border: '1px solid lightgray',
+                        borderRadius: '4px',
+                    },
+                },
+                chooseButton: {
+                    style: { backgroundColor: 'aliceblue', color: 'black' },
+                },
+                content: ({ props, state }) => {
+                    const done = [0, 100].includes(state.progress);
+                    return {
+                        style: { display: done ? 'none' : '' },
+                    };
+                },
+                input: { ariaLabel: $gettext('Upload an image') },
+            }"
+            @before-send="addHeader($event)"
+            @upload="upload($event)"
+            @error="showError($event)"
+        />
+        <div class="images">
+            <ImageEditor
+                v-for="image in item.images"
+                :key="image.id"
+                :image="image"
+            />
+            <p v-if="!item.images.length">
+                {{ $gettext("No images.") }}
+            </p>
+        </div>
+    </div>
+</template>
+
+<style scoped>
+.images-container {
+    margin: 1rem 1rem 3rem 1rem;
+    display: flex;
+    flex-direction: column;
+    width: 100%;
+}
+
+.images {
+    margin-top: 1.5rem;
+    gap: 1.5rem;
+}
+
+h4 {
+    color: v-bind(ARCHES_CHROME_BLUE);
+    margin-top: 0;
+    font-size: 1.33rem;
+}
+
+p {
+    font-size: 1.2rem;
+}
+
+.images {
+    display: flex;
+    flex-direction: column;
+    gap: 3rem;
+}
+
+:deep(input[type="file"]) {
+    /* override arches.css */
+    /* PrimeVue uses a hidden input for screen readers */
+    display: none;
+}
+
+:deep(.p-fileupload-content:empty) {
+    display: none;
+}
+</style>
